+++ src/Common/Builder.fs
@@ -81,9 +81,15 @@ module ApplicationBuilder =
                     Resource = ResourceAvailability.createFromStrings "graylog" graylogService graylog Audience.Sys
                     Interval = 30<KafkaApplication.Second>
                     Checker = fun () ->
-                        if graylogService |> Graylog.Diagnostics.isAlive |> Async.RunSynchronously
-                        then Up
-                        else Down
+                        graylogService
+                        |> Graylog.Diagnostics.isAliveResult
+                        |> Async.RunSynchronously
+                        |> function
+                            | Ok isAlive when isAlive -> Up
+                            | Ok _ -> Down
+                            | Error e ->
+                                parts.Logger.Error "Graylog Resource" <| sprintf "%A" e
+                                Down
                 }
 
                 return { parts